refactor: Player가 PlayerInfo가 아닌 PlayerBet를 인스턴스 필드로 가지도록 변경한다
